moved print_JFH_param_curve() from RPOD.py to JetFiringHistory.py
