+++ pyrpod/RPOD.py
@@ -2,7 +2,6 @@ import numpy as np
 import pandas as pd
 import os
 import math
-import sympy as sp
 import matplotlib.pyplot as plt
 from mpl_toolkits import mplot3d
 import configparser
@@ -12,27 +11,8 @@ from stl import mesh
 from pyrpod.LogisticsModule import LogisticsModule
 from pyrpod.MissionPlanner import MissionPlanner
 
-from pyrpod.file_print import print_JFH
+
 from tqdm import tqdm
-
-# Helper functions
-def rotation_matrix_from_vectors(vec1, vec2):
-    """ Find the rotation matrix that aligns vec1 to vec2
-    :param vec1: A 3d "source" vector
-    :param vec2: A 3d "destination" vector
-    :return mat: A transform matrix (3x3) which when applied to vec1, aligns it with vec2.
-    """
-    a, b = (vec1 / np.linalg.norm(vec1)).reshape(3), (vec2 / np.linalg.norm(vec2)).reshape(3)
-    v = np.cross(a, b)
-    c = np.dot(a, b)
-    s = np.linalg.norm(v)
-    kmat = np.array([[0, -v[2], v[1]], [v[2], 0, -v[0]], [-v[1], v[0], 0]])
-    rotation_matrix = np.eye(3) + kmat + kmat.dot(kmat) * ((1 - c) / (s ** 2))
-    return rotation_matrix
-
-def make_norm(vector_value_function):
-    """Calculate vector norm/magnitude using the Pythagoream Theorem."""
-    return sp.sqrt(sp.Pow(vector_value_function[0],2) + sp.Pow(vector_value_function[1],2))
 
 class RPOD (MissionPlanner):
     """
@@ -536,214 +516,26 @@ class RPOD (MissionPlanner):
             self.target.convert_stl_to_vtk_strikes(path_to_vtk, cellData, target)
             # print()
 
+    # def print_jfh_1d_approach(self):
+    #     """
+    #         Method creates JFH data for axial approach using simpified physics calculations.
 
-    def graph_param_curve(self, t, r_of_t):
-        ''' Used to quickly prototype and visualize a proposed approach path.
-            Calculates the unit tangent vector at a given timestep and 
-            rotates the STL file accordingly. Data is plotted using matlab
+    #         NOTE: Methods does not take any parameters. It assumes that self.case_dir
+    #         and self.config are instatiated correctly. Potential defensive programming statements?
 
-            Current method is old and needs updating.
+    #         Parameters
+    #         ----------
+    #         None
 
-            Parameters
-            ----------
-            t : sp.symbol
-                Time (t) is the independent variable used to evaulte the position vector equation.
+    #         Returns
+    #         -------
+    #         Method doesn't currently return anything. Simply sets class members as needed.
+    #         Does the method need to return a status message? or pass similar data?
 
-            r_of_t : list<expressions?>
-                List containing position vector expression for trajectory.
-                X/Y/Z positions are de-coupled and only dependent on time.
-
-            Returns
-            -------
-            Method doesn't currently return anything. Simply sets class members as needed.
-            Does the method need to return a status message? or pass similar data?
-        '''
-
-        t_values = np.linspace(0,2*np.pi,50)
-
-        # Symbolic Calculations of tangent and normal unit vectors
-        r = r_of_t
-        rprime = [diff(r[0],t), diff(r[1],t), diff(r[2],t)]
-        tanvector = [rprime[0]/make_norm(rprime), rprime[1]/make_norm(rprime), rprime[2]/make_norm(rprime)]
-        tanprime = [diff(tanvector[0],t), diff(tanvector[1],t), diff(tanvector[2],t)]
-        normalvector = [tanprime[0]/make_norm(tanprime), tanprime[1]/make_norm(tanprime), tanprime[1]/make_norm(tanprime)]
-        tan_vector_functions = [lambdify(t, tanvector[0]),lambdify(t, tanvector[1]), lambdify(t, tanvector[2])]
-        normal_vector_functions = [lambdify(t, normalvector[0]),lambdify(t, normalvector[1]), lambdify(t, normalvector[2])]
-        value_functions = [lambdify(t, r[0]), lambdify(t, r[1]), lambdify(t, r[2])]
-
-        # Save data of evaluated position and velocity functions. 
-        x, y, z = [value_functions[0](t_values), value_functions[1](t_values), value_functions[2](t_values)]
-        dx, dy, dz = [tan_vector_functions[0](t_values), tan_vector_functions[1](t_values), tan_vector_functions[2](t_values)]
-
-        # draw the vectors along the curve and Graph STL.
-        for i in range(len(t_values)):
-            # Graph path
-            # ax = plt.figure().add_subplot(projection='3d')
-            figure = plt.figure()
-            ax = mplot3d.Axes3D(figure)
-            ax.plot(x, y, z, label='position curve')
-            ax.legend()
-            normal_location = t_values[i]
-
-            # Load, Transform, and Graph STL
-            VV = mesh.Mesh.from_file('../stl/cylinder.stl')
-            VV.points = 0.2 * VV.points
-
-            r = [x[i], y[i], z[i]]
-            dr = [dx, dy, dz[i]]
-
-            # Calculate require rotation matrix from initial orientation.
-            x1 = [1, 0, 0]
-            rot = np.matrix(rotation_matrix_from_vectors(x1, dr))
-
-            VV.rotate_using_matrix(rot.transpose())
-            VV.translate(r)
-
-            ax.add_collection3d(
-                mplot3d.art3d.Poly3DCollection(VV.vectors)
-            )
-
-            # print(
-            # tan_vector_functions[0](normal_location),
-            # tan_vector_functions[1](normal_location),
-            # tan_vector_functions[2](normal_location)
-            # )
-            # print()
-            length = 1
-            ax.quiver(
-                value_functions[0](normal_location),
-                value_functions[1](normal_location),
-                value_functions[2](normal_location),
-                tan_vector_functions[0](normal_location),
-                tan_vector_functions[1](normal_location),
-                tan_vector_functions[2](normal_location),
-                color='g',
-                length = length 
-            )
-
-            # ax.quiver(
-            #     value_functions[0](normal_location),
-            #     value_functions[1](normal_location),
-            #     value_functions[2](normal_location),
-            #     normal_vector_functions[0](normal_location),
-            #     normal_vector_functions[1](normal_location),
-            #     normal_vector_functions[2](normal_location),
-            #     color='r'
-            # )
-            print(i)
-
-            ax.set_xlabel('X')
-            ax.set_ylabel('Y')
-            ax.set_zlabel('Z')
-
-            if i < 10:
-                index = '00' + str(i)
-            elif i < 100:
-                index = '0' + str(i)
-            else:
-                index = str(i)
-
-            plt.savefig('img/observer-a-' + str(index) + '.png')
-
-            plt.close() 
-
-    def print_JFH_param_curve(self, jfh_path, t, r_of_t, align = False):
-        ''' Used to produce JFH data for a proposed approach path.
-            Calculates the unit tangent vector at a given timestep and DCMs for
-            STL file rotations. Data is then saved to a text file.
+    #     """
 
 
-            Parameters
-            ----------
-            jfh_path : str
-                Path to file for saving JFH data.
 
-            t : sp.symbol
-                Time (t) is the independent variable used to evaulte the position vector equation.
-
-            r_of_t : list<expressions?>
-                List containing position vector expression for trajectory.
-                X/Y/Z positions are de-coupled and only dependent on time.
-
-            aligh : Boolean
-                Determines whether or not STL surface is rotated according to unit tangent vector.
-
-            Returns
-            -------
-            Method doesn't currently return anything. Simply sets class members as needed.
-            Does the method need to return a status message? or pass similar data?
-        '''
-
-        # t_values = np.linspace(0,2*np.pi,100)
-        t_values = np.linspace(0, 50, 20)
-
-        # Symbolic Calculations of tangent and normal unit vectors
-        r = r_of_t
-        rprime = [sp.diff(r[0],t), sp.diff(r[1],t), sp.diff(r[2],t)]
-        tanvector = [rprime[0]/make_norm(rprime), rprime[1]/make_norm(rprime), rprime[2]/make_norm(rprime)]
-        tanprime = [sp.diff(tanvector[0],t), sp.diff(tanvector[1],t), sp.diff(tanvector[2],t)]
-        normalvector = [tanprime[0]/make_norm(tanprime), tanprime[1]/make_norm(tanprime), tanprime[1]/make_norm(tanprime)]
-        tan_vector_functions = [sp.lambdify(t, tanvector[0]),sp.lambdify(t, tanvector[1]), sp.lambdify(t, tanvector[2])]
-        normal_vector_functions = [sp.lambdify(t, normalvector[0]),sp.lambdify(t, normalvector[1]), sp.lambdify(t, normalvector[2])]
-        value_functions = [sp.lambdify(t, r[0]), sp.lambdify(t, r[1]), sp.lambdify(t, r[2])]
-
-        # Save data of evaluated position and velocity functions. 
-        x, y, z = [value_functions[0](t_values), value_functions[1](t_values), value_functions[2](t_values)]
-        dx, dy, dz = [tan_vector_functions[0](t_values), tan_vector_functions[1](t_values), tan_vector_functions[2](t_values)]
-
-        # print(type(dx), type(dy), type(dz))
-
-        # print(dx.size, dy.size, dz.size)
-
-        # When derivatives reduce to constant value the lambda function will reutrn a float 
-        # instead of np.array. These if statements are here fill an array with that float value.
-        # print(type(dx), dx) 
-        # print(type(dy), dy) 
-        if type(x) == int:
-            x = np.full(t_values.size, x)
-
-        if type(dx) == int or dx.size == 1:
-            # print('dx is contant')
-            # print(dx)
-            dx = np.full(t_values.size, dx)
-            # x = np.full(t_values.size, )
-
-        if type(y) == int:
-            y = np.full(t_values.size, y)
-        if type(dy) == int or dy.size == 1:
-            # print('dy is contant')
-            # print(dy)
-            dy = np.full(t_values.size, dy)
-        
-        if type(z) == int:
-            z = np.full(t_values.size, z)
-        if type(dz) == int or dz.size == 1:
-            # print('dz is contant')
-            # print(dz)
-            dz = np.full(t_values.size, dz)
-
-        # print(type(dx), type(dy), type(dz))
-        # print(type(x), type(y), type(z))
-
-        # Save rotation matrix for each time step
-        rot = []
-        if align:
-            for i in range(len(t_values)):
-                dr = [dx[i], dy[i], dz[i]]
-
-                # Calculate required rotation matrix from initial orientation.
-                x1 = [1, 0, 0]
-                rot.append(np.matrix(rotation_matrix_from_vectors(x1, dr)))
-        else:
-            for i in range(len(t_values)):
-                # Calculate required rotation matrix from initial orientation.
-                y1 = [0, 0, -1]
-                x1 = [1, 0, 0]
-                rot.append(np.matrix(rotation_matrix_from_vectors(x1, y1)))
-
-        r = [x, y, z]
-        # dr = [dx, dy, dz]
-        print_JFH(t_values, r,  rot, jfh_path)
 
 def make_test_jfh():
